fix(browser): destroy webview in IPC close path to prevent background video leak
Extract webview registry and destroyPersistentWebview into a side-effect-free
module so useIpcEvents.ts can import it without pulling in BrowserPane's
drag-listener side effects. The onCloseActiveTab IPC handler (triggered when
Cmd+W is pressed inside a webview guest) was missing the destroy call, leaving
parked webviews that resume paused media.
