+++ src/renderer/src/components/browser-pane/webview-registry.ts
@@ -1,0 +1,21 @@
+import { clearLiveBrowserUrl } from './browser-runtime'
+
+export const webviewRegistry = new Map<string, Electron.WebviewTag>()
+export const registeredWebContentsIds = new Map<string, number>()
+export const parkedAtByTabId = new Map<string, number>()
+
+export function destroyPersistentWebview(browserTabId: string): void {
+  const webview = webviewRegistry.get(browserTabId)
+  if (!webview) {
+    registeredWebContentsIds.delete(browserTabId)
+    parkedAtByTabId.delete(browserTabId)
+    clearLiveBrowserUrl(browserTabId)
+    return
+  }
+  void window.api.browser.unregisterGuest({ browserPageId: browserTabId })
+  webview.remove()
+  webviewRegistry.delete(browserTabId)
+  registeredWebContentsIds.delete(browserTabId)
+  parkedAtByTabId.delete(browserTabId)
+  clearLiveBrowserUrl(browserTabId)
+}
